# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.AuthResponse:
    properties:
      message:
        description: Сообщение о статусе авторизации
        example: Authentication successful
        type: string
      token:
        description: JWT токен для последующих запросов
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.UserResponse'
        description: Детали пользователя
    type: object
...
  models.LoginRequest:
    properties:
      email:
        description: Email пользователя
        example: admin@example.com
        type: string
      password:
        description: Пароль пользователя
        example: admin_hash_secure
        type: string
    type: object
...
  models.RequestCodeRequest:
    properties:
      email:
        description: Email пользователя, запрашивающего код
        example: observer1@example.com
        type: string
      event_id:
        description: ID мероприятия, к которому запрашивается доступ
        example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        type: string
    type: object
...
  models.UserResponse:
    properties:
      email:
        description: Email пользователя
        example: admin@example.com
        type: string
      id:
        description: ID пользователя
        example: 1
        type: integer
      name:
        description: Имя пользователя
        example: Главный Администратор
        type: string
      role:
        description: Роль пользователя (например, "admin" или "observer")
        example: admin
        type: string
      type_of_activity:
        description: Тип деятельности
        example: IT
        type: string
    type: object
  models.VerifyCodeRequest:
    properties:
      code:
        description: Одноразовый код, полученный по почте
        example: "123456"
        type: string
      email:
        description: Email пользователя
        example: observer1@example.com
        type: string
      event_id:
        description: ID мероприятия
        example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Сервис для управления событиями, картами и точками интереса.
  title: Event Designer API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Обрабатывает вход администратора по почте и паролю. Возвращает
        JWT-токен с ролью "admin".
      parameters:
      - description: Учетные данные администратора
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный вход, возвращает токен
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Invalid JSON
          schema:
            type: string
        "401":
          description: Invalid email or password / Only admins can log in with password
          schema:
            type: string
        "500":
          description: Database error / Failed to generate token
          schema:
            type: string
      summary: Вход администратора
      tags:
      - Auth
  /auth/request-code:
    post:
      consumes:
      - application/json
      description: Отправляет одноразовый код потверждения на email для входа на мероприятие.
        Требуется, чтобы email был разрешен для EventID.
      parameters:
      - description: Event ID и Email участника
        in: body
        name: requestCode
        required: true
        schema:
          $ref: '#/definitions/models.RequestCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Код успешно сгенерирован и отправлен
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid JSON / Required fields missing / Invalid event_id format
          schema:
            type: string
        "403":
          description: Email not allowed for this event
          schema:
            type: string
        "500":
          description: Database error / Failed to generate code / Failed to save code
          schema:
            type: string
      summary: Запрос кода доступа
      tags:
      - Auth
  /auth/verify-code:
    post:
      consumes:
      - application/json
      description: Потверждает код доступа, предоставленный по email. При успешной
        проверке возвращает JWT-токен для доступа к ресурсам мероприятия. Если пользователь
        не существует, он будет создан как "observer".
      parameters:
      - description: Event ID, Email и полученный код доступа
        in: body
        name: verifyCode
        required: true
        schema:
          $ref: '#/definitions/models.VerifyCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная аутентификация, возвращает токен
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Invalid JSON / Required fields missing / Invalid event_id format
          schema:
            type: string
        "401":
          description: Invalid or expired code
          schema:
            type: string
        "403":
          description: Email not allowed for this event
          schema:
            type: string
        "500":
          description: Database error / Failed to generate token
          schema:
            type: string
      summary: Подтверждение кода и вход участника
      tags:
      - Auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"